Move shortcuts to DllManipulatorEditor
