+++ scripts/DllManipulatorScript.cs
@@ -4,7 +4,6 @@ using System.Threading;
 using System.Linq;
 using UnityEngine;
 using UnityNativeTool.Internal;
-using UnityEditor.ShortcutManagement;
 
 namespace UnityNativeTool
 {
@@ -84,26 +83,6 @@ namespace UnityNativeTool
 
             timer.Stop();
             InitializationTime = timer.Elapsed;
-        }
-
-        #if UNITY_2019_1_OR_NEWER
-        [Shortcut("Tools/Load All Dlls", KeyCode.D, ShortcutModifiers.Alt | ShortcutModifiers.Shift)]
-        #else
-        [MenuItem("Tools/Load All Dlls #&d")]
-        #endif
-        public static void LoadAllShortcut()
-        {
-            DllManipulator.LoadAll();
-        }
-
-        #if UNITY_2019_1_OR_NEWER
-        [Shortcut("Tools/Unload All Dlls", KeyCode.D, ShortcutModifiers.Alt)]
-        #else
-        [MenuItem("Tools/Unload All Dlls &d")]
-        #endif
-        public static void UnloadAll()
-        {
-            DllManipulator.UnloadAll();
         }
 
         private void OnDestroy()
